+++ plotpy/items/shape/__init__.py
@@ -9,7 +9,7 @@ from .ellipse import EllipseShape
 from .marker import Marker
 from .point import PointShape
 from .polygon import PolygonShape
-from .range import XRangeSelection
+from .range import XRangeSelection, YRangeSelection
 from .rectangle import ObliqueRectangleShape, RectangleShape
 from .segment import SegmentShape
 from .svg import CircleSVGShape, RectangleSVGShape, SquareSVGShape